separate file for dice node inner details
